==> src/01_filtering_mj.ipynb <==
library(readxl)
library(readr)

#Example of filtering
#filtered = data.frame(tconst=c(1,2,3,4,10), att2=c("1","2","3","4","5"))
#new = data.frame(tconst=c(2,3,4,4,3,7,8,1), att3=c("a","b","c","d","e","f","g","h"))
filtered = data.frame(tconst=c(2,3,4,4,3,7,8,1), att3=c("a","b","c","d","e","f","g","h"))
new = data.frame(tconst=c(1,2,3,4,4,10), att2=c("1","2","3","4","4","5"))
a = dplyr::filter(new, new$tconst %in% filtered$tconst)
a

### title akas

title_akas = read.delim("./../dataset/TitleAkas.tsv",header=TRUE)
filtered_title_basics = read.delim("./../dataset/filtered.title.basics.tsv",header=TRUE)

title_akas_filtered = dplyr::filter(title_akas, title_akas$titleId %in% filtered_title_basics$tconst)

write_tsv(title_akas_filtered,'filtered_title_akas.tsv')

### title principals

title_principals = read.delim("./../dataset/TitlePrincipals.tsv",header=TRUE)

title_principals_filtered = dplyr::filter(title_principals, title_principals$tconst %in% filtered_title_basics$tconst)

write_tsv(title_principals_filtered, "filtered_title_principals.tsv")

### name basics

name_basics = read.delim("./../dataset/NameBasics.tsv",header=TRUE)
name_basics_filtered = dplyr::filter(name_basics, name_basics$nconst %in% title_principals_filtered$nconst)
write_tsv(name_basics_filtered,'filtered_name_basics.tsv')

### title crew

title_crew = read.delim("./../dataset/TitleCrew.tsv",header=TRUE)
title_crew_filtered = dplyr::filter(title_crew, title_crew$tconst %in% filtered_title_basics$tconst)
write_tsv(title_crew_filtered,'filtered_title_crew.tsv')

### title ratings

title_ratings = read.delim("./../dataset/TitleRatings.tsv",header=TRUE)
title_ratings_filtered = dplyr::filter(title_ratings, title_ratings$tconst %in% filtered_title_basics$tconst)
write_tsv(title_ratings_filtered,'filtered_title_ratings.tsv')
