==> src/01_title.crew - filtering process.ipynb <==
library(readxl)

new <- read.delim('/home/fatimab/Documents/FEUP/Visualização_de_Dados/VD_project/dataset/title.principals.tsv',header=TRUE,nrows=)
print("New loaded")
filtered <- read.delim("/home/fatimab/Documents/FEUP/Visualização_de_Dados/VD_project/dataset/filtered.title.basics.tsv",header=TRUE)
print("Filtered Title basics loaded")

df = dplyr::filter(new, new$tconst %in% filtered$tconst)

print("Saving new filtered...")
write.table(df,file='filtered.title.principals.tsv',quote=FALSE,sep='\t',row.names=FALSE)
print("Done!")
