==> src/02_cleaning_title_akas_bv.ipynb <==
## Arranging Title Akas

#tb=read.delim('cleaned_filtered_title_basics.tsv',header=TRUE)

#head(tb)

ta=read.delim('../../IMDb_Datasets_Filtered/filtered_title_akas.tsv',header=TRUE)

ta$ordering = NULL

nrow(ta[ta$attributes=='\\N',])/nrow(ta) #95.1%

ta$attributes = NULL

nrow(ta[ta$types=='\\N',])/nrow(ta) #52.4%

ta$types = NULL

nrow(ta[ta$language=='\\N',])/nrow(ta) #87.1%

ta$language = NULL

ta$title = NULL

ta_clean=ta #Further changes will take place in the ta_clean dataframe

head(ta_clean)

nrow(ta_clean[ta_clean$region=='\\N',])/nrow(ta) #13.6% of missing values regarding the region attribute

ta_clean=subset(ta_clean,region!='\\N') #Let's remove them

ta_clean[is.na(ta_clean$titleId),] #They're gone (properly gone!)

nrow(ta_clean[ta_clean$isOriginalTitle=='\\N',])/nrow(ta) #0.0016% are missing

ta_clean=subset(ta_clean,isOriginalTitle!='\\N') #Let's remove them

ta_clean[is.na(ta_clean$titleId),] #They're gone (properly gone!)

nrow(ta_clean)

library(readr)
write_tsv(ta_clean,'../../cleaned_filtered_title_akas.tsv')

#### (SOLVED!) If one tries to remove columns something weird happens... We obtain NA values on the titleId variable... I don't know why it happens

## Experiment

#titleid=unique(ta$titleId) #All titleId's

#length(titleid) #Number of different movies

#The for cycle never stops
a=data.frame(titleid[1:length(titleid)])
regions=c()
original_tit=c()

for (i in 1:length(titleid)){
    region=ta_clean[ta_clean$titleId==titleid[i],]$region
    orig_tit=ta_clean[ta_clean$titleId==titleid[i],]$isOriginalTitle
    regions=append(regions,list(region))
    original_tit=append(original_tit,list(orig_tit))
}

a$regions=regions
a$titleOriginal=original_tit
