==> src/03_eda_namebasics11_mj.ipynb <==
### Imports

library(ggplot2)
library(base)

### Read File

path = "./../dataset/"
nb <- read.delim(paste(path, "NameBasics11.tsv", sep=""), header=TRUE)

head(nb)

str(nb)

summary(nb)

summary(nb$primaryName)

str(unique(nb$primaryName))

summary(nb$primaryProfession)

str(unique(nb$primaryProfession))

sum(nb$deathYear=="\\N")

a = unlist(strsplit(as.character(nb$primaryProfession[1]), ","),'[.]')
a[1]
a[2]
a[3]
class(a)
