==> src/03_Merging_bv.ipynb <==
tc=read.delim('IMDb_Datasets_Filtered/cleaned_filtered_title_crew.tsv',header=TRUE)

tb=read.delim('cleaned_filtered_title_basics.tsv',header=TRUE)

tr=read.delim('IMDb_Datasets_Filtered/filtered_title_ratings.tsv',header=TRUE)

head(tb)

tbtr=merge(tb,tr, by="tconst",all.x=TRUE) #Merging title_basics with title_ratings
#The all.x = True means that the movies which are in title_ratings but are not in title_basics won't be in tbtr

nrow(tbtr)
nrow(tb)
#As expected, the number of rows is the same

head(tbtr)

tbtrtc=merge(tbtr,tc,by='tconst',all.x=TRUE) #Merging tbtr with title_crew
#All movies which are in title_crew but not in title_basics will not be merged

nrow(tbtrtc) #As expected

head(tbtrtc)
tail(tbtrtc)

tbtrtc[is.na(tbtrtc$averageRating),9]=-1 #Replacing the missing averageRatings by -1

tbtrtc[is.na(tbtrtc$numVotes),10]=-1 #Replacing the missing numVotes by -1

tbtrtc[is.na(tbtrtc$n_directors),11]=-1 #Replacing the missing n_directors by -1

tbtrtc[is.na(tbtrtc$n_writers),12]=-1 #Replacing the missing n_writers by -1

library(readr)
write_tsv(tbtrtc,'merge_basics_ratings_crew.tsv')
