==> src/04_boxplot_titles_per_person.ipynb <==
nb=read.delim('./../dataset/cleaned_filtered_name_basics.tsv')

merge=read.delim('./../dataset/merge_basics_ratings_crew.tsv')

head(nb)

head(merge)

titles1=nb[,6,drop=F]
titles2=nb[,7,drop=F]
titles3=nb[,8,drop=F]
titles4=nb[,9,drop=F]
colnames(titles1)='tconst'
colnames(titles2)='tconst'
colnames(titles3)='tconst'
colnames(titles4)='tconst'

titles1[(titles1$tconst=="0"),]= "NA"
titles2[(titles2$tconst=="0"),] = "NA"
titles3[(titles3$tconst=="0"),] = "NA"
titles4[(titles4$tconst=="0"),] = "NA"

ratings=merge[,c(1,9)]

summary(ratings)

ratings = ratings[complete.cases(ratings),]

tail(titles4)

summary(ratings)

nrow(titles1)

library(tidyverse)

titles_1_join = left_join(titles1,ratings,by="tconst")
titles_2_join = left_join(titles2,ratings,by="tconst")
titles_3_join = left_join(titles3,ratings,by="tconst")
titles_4_join = left_join(titles4,ratings,by="tconst")

tail(titles4)

tail(titles_4_join)

titles=cbind(titles_1_join,titles_2_join,titles_3_join,titles_4_join)

head(titles)

tail(titles)

nrow(titles)

ratings_titles=titles[,c(2,4,6,8)]

row_means = rowMeans(ratings_titles, na.rm = TRUE,dims = 1)

head(row_means)

summary(row_means)

#remove NAs
row_means = row_means[complete.cases(row_means)]

summary(row_means)

library(ggplot2)

row_means_df=data.frame(row_means)

head(row_means_df)

g = ggplot(row_means_df,aes(x="",y=row_means)) + geom_violin(fill="skyblue",color="skyblue4",alpha=0.3)+
geom_boxplot(width=0.4,fill="white",color="skyblue4") +
theme_bw(base_size = 15)+
scale_y_continuous(breaks=seq(0,10,1))+
coord_flip()+
theme(axis.title.y=element_blank(),axis.text.y=element_blank(),
axis.ticks.y=element_blank(),legend.position="none",plot.title=element_text(
hjust=0.5,size=30)) +
labs(y=expression("Mean Average Ratings"),x="")

ggsave(plot=g,file="average_ratings_per_person.png",width=3,height=3,dpi=300)
