==> src/05_hypothesis_people_and_ratings.ipynb <==
## hypothesis success: people and ratings

Grafo: vértices->pessoas; 

aresta entre duas pessoas caso tenham trabalhado juntas; 

a espessura da aresta é tanto maior quanto maior for o nº de filmes em comum 

# filtered name basics containing only actors and actresses
names = read.csv('./../dataset/name_basics_actors_actresses.csv',header=TRUE)

head(names)

df = data.frame(nconst=factor(), knownForTitles=factor())

head(df)

#df$nconst = rep(names$nconst, 4)

#head(df)

merged = read.delim('./../dataset/merge_basics_ratings_crew.tsv',header=TRUE)

head(merged)
